--- a/ui/package.json
+++ b/ui/package.json
@@ -111,7 +111,6 @@
     "eslint-plugin-import": "^2.16.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^3.0.1",
-    "happypack": "^5.0.1",
     "html-loader": "^0.5.5",
     "html-minifier": "^3.5.21",
     "html-minifier-loader": "^1.4.1",