- [x] Upgrading webpack to v4
- [x] Processing time has been reduced from 123secs to 34sec on production and from 40secs to 34secs on develepment

_Which was the motivation of this change?_
My OS crashed from time to time when processing using Webpack before this.


Benchmarks (/usr/bin/time npx webpack --silent):
- Before:
  - production + happypack
145.17user 3.45system **1:23.42elapsed** 178%CPU (0avgtext+0avgdata 1247288maxresident)k
3304inputs+44432outputs (0major+1284098minor)pagefaults 0swaps
  - development + happypack
96.99user 2.77system **0:40.61elapsed** 245%CPU (0avgtext+0avgdata 1100384maxresident)k
17608inputs+82000outputs (0major+715383minor)pagefaults 0swaps
- After:
  - production + webpack 4 + thread-loader
80.69user 2.83system **0:34.03elapsed** 245%CPU (0avgtext+0avgdata 661888maxresident)k
304inputs+74440outputs (0major+569763minor)pagefaults 0swaps
  - development + webpack 4 + style-loader
42.78user 1.04system **0:34.76elapsed** 126%CPU (0avgtext+0avgdata 644076maxresident)k
0inputs+91208outputs (0major+220049minor)pagefaults 0swaps